Initial separation of react component from render component
